requirements.txt
openai
mcp
uvicorn
python-dotenv
python-multipart
pydantic
pillow
httpx
tenacity
pytest
pytest-asyncio
